rename searching components to search components
