+++ web-ui/src/components/course/course-list/course-list.tsx
@@ -11,7 +11,7 @@ import {
   emitEvent,
   EVENT_ERROR_RESPONSE,
 } from '@/matomo';
-import { SearchResultLoading } from '@/components/searching/search-result-loading/search-result-loading';
+import { SearchResultLoading } from '@/components/search/search-result-loading/search-result-loading';
 
 const { Search } = Input;
 const { Link } = Typography;
