+++ web-ui/src/components/search/search-by-image/search-by-image.less
