+++ web-ui/src/components/search/search-by-image/search-by-image.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import apiClient, { ServerErrorResponse, ServerResponse } from '@/http';
 import { LoadingOutlined, ReloadOutlined } from '@ant-design/icons';
-import { SearchResultLoading } from '@/components/searching/search-result-loading/search-result-loading';
+import { SearchResultLoading } from '@/components/search/search-result-loading/search-result-loading';
 import { history } from 'umi';
 import { LecturePreviewCompact } from '@/components/lecture/lecture-preview/lecture-preview';
 import { Highlight, Lecture } from '@/types/lecture';
