+++ web-ui/src/components/search/search-huge/search-huge.less
