+++ web-ui/src/components/search/search-result-loading/search-result-loading.less
