+++ web-ui/src/components/search/search-results/search-result.less
